requirements.txt
# requirements.txt
quart
aiobotocore
azure-identity
azure-mgmt-resource
azure-mgmt-compute
python-dotenv
aiohttp
hypercorn
asgiref
gunicorn
uvicorn
pytest
pytest-asyncio
Werkzeug